% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/telephone.R
\name{telephone}
\alias{telephone}
\title{Run telephone game}
\usage{
telephone(word)
}
\arguments{
\item{word}{word to pass to the first player}
}
\value{
vector of length n_players + 1, first element is the input word, each subsequent element is a word selected by a player.
}
\description{
Run one iteration of the telephone game
}
